Add minimal ECS framework
